backend/requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
torch>=2.2.0
transformers>=4.11.3
PyMuPDF>=1.18.17
Pillow>=8.3.2
python-multipart>=0.0.5 